+++ src/main/java/br/com/maria/imposto/renda/ImpostoDeRendaIsento.java
@@ -1,4 +1,4 @@
-package renda;
+package main.java.br.com.maria.imposto.renda;
 
 import java.math.BigDecimal;
